Migrate to React Router v7: Fix imports, routes, and configuration
